+++ app/page.tsx
@@ -3,15 +3,14 @@ import About from "@/components/sections/About";
 import Experience from "@/components/sections/Experience";
 import Projects from "@/components/sections/Projects";
 import Contact from "@/components/sections/Contact";
-
-import Skills from "@/components/subsections/Skills";
-import Hobbies from "@/components/subsections/Hobbies";
+import Skills from "@/components/sections/Skills";
 
 import BackToTop from "@/components/BackToTop";
 
 const Home = () => {
+
   return (
-    <div> 
+    <div>
       <Hero />
       <BackToTop />
       <div className="flex flex-col mb-12" id="mainContent">