+++ components/Footer.tsx
@@ -7,7 +7,7 @@ const Footer = async () => {
       id="footer"
       className="bg-slate-600 text-2xl flex flex-col w-full justify-center items-center p-8 space-y-2"
     >
-      <div className="flex flex-col space-y-2 md:space-y-0 md:space-x-4 md:flex-row items-center justify-center w-full">
+      <div className="flex flex-col space-y-1 md:space-y-0 md:space-x-4 md:flex-row items-center justify-center w-full">
         <div className="flex flex-row justify-center space-x-4">
           <Link href="#about" className="footer-link">
             About
@@ -26,18 +26,9 @@ const Footer = async () => {
         </div>
       </div>
       <div className="flex flex-col space-y-1 text-center">
-        <p>
-          Copyright &copy; {new Date().getFullYear()}{" "}
-          <Link
-            href="https://github.com/p55d2k"
-            target="_blank"
-            className="text-gray-400 hover:text-gray-300 transition-all ease-in-out duration-300"
-          >
-            p55d2k
-          </Link>{" "}
-          (Zhi Kai)
+        <p className="text-sm">
+          Copyright &copy; {new Date().getFullYear()} p55d2k
         </p>
-        <p className="text-sm">This was created in 2024, when I was 14! 🎉</p>
         <p className="text-sm">
           <Link
             href="https://github.com/p55d2k/zk-website"