+++ components/cards/ProjectCard.tsx
@@ -4,9 +4,10 @@ import Link from "next/link";
 
 interface ProjectCardProps {
   proj: Project;
+  whiteText?: boolean;
 }
 
-const ProjectCard = ({ proj }: ProjectCardProps) => {
+const ProjectCard = ({ proj, whiteText }: ProjectCardProps) => {
   return (
     <Link
       href={`/project/${proj.name.toLowerCase().replaceAll(" ", "_")}`}
@@ -17,13 +18,15 @@ const ProjectCard = ({ proj }: ProjectCardProps) => {
           src={proj.screenshot}
           alt={proj.name}
           width={400}
-          height={400}
-          className="bg-white w-full border-2 border-white rounded-lg"
+          height={200}
+          className="bg-white border-2 border-white w-full max-h-[300px]"
         />
         <div className="text-center bg-gradient-to-r from-pink-500 via-purple-500 to-blue-500 text-opacity-0 pb-3 text-white bg-clip-text font-extrabold text-4xl pt-3">
           {proj.name}
         </div>
-        <p className="text-center text-sm text-amber-500">{proj.description}</p>
+        <p className={`text-center text-sm ${!whiteText && "text-amber-500"}`}>
+          {proj.description}
+        </p>
       </div>
     </Link>
   );