+++ components/sections/About.tsx
@@ -11,10 +11,10 @@ const About = () => {
   const flexSection = useMedia("(max-width: 900px)", false);
 
   return (
-    <div className="md:pt-20 flex flex-col" id="about">
-      <div className="flex flex-col text-white px-8 sm:px-16 md:px-20 space-y-6 items-center">
+    <div className="flex flex-col" id="about">
+      <div className="flex flex-col text-white px-8 sm:px-16 md:px-20 space-y-6 items-center justify-center min-h-screen">
         <SectionHeader sequence={["about me", 4000, "自我介绍", 4000]}>
-          i am <span className="text-blue-500">wan zhi kai</span>
+          i am <span className="text-red-500">wan zhi kai</span>
         </SectionHeader>
         <div
           className={`flex ${
@@ -23,85 +23,52 @@ const About = () => {
               : "flex-row space-x-6 lg:space-x-14 xl:space-x-20"
           } items-center justify-center py-3`}
         >
-          <div className="hidden sm:flex flex-col space-y-1">
-            <Image src="/imgs/zk2.jpeg" alt="" width={400} height={400} />
-            <div className="text-center italic">
-              Me riding a camel in Hatta (حتا), Dubai, UAE 2023
-            </div>
-          </div>
-          <div className="flex flex-col space-y-1 max-w-[500px]">
-            <h3 className="text-2xl md:text-3xl text-center md:text-left">
+          <Image src="/imgs/zk2.jpeg" alt="" width={500} height={500} />
+          <div
+            className={`flex flex-col space-y-1 ${
+              flexSection
+                ? "max-w-[500px]"
+                : "max-w-[400px] lg:max-w-[475px] xl:max-w-[550px]"
+            }`}
+          >
+            <h3 className="text-2xl md:text-3xl lg:text-5xl xl:text-6xl text-center md:text-left">
               Hi there! 👋
             </h3>
-            <div className="flex flex-col space-y-3 text-sm md:text-base text-left font-light pt-2">
+            <div className="flex flex-col space-y-3 text-sm md:text-base xl:text-lg text-left font-light pt-2">
               <p>
-                I am <span className="text-blue-500">Wan Zhi Kai</span>, a
-                software engineer, specialised in frontend development, based in
-                Singapore. I am currently a high school student at{" "}
-                <Link
-                  className="text-red-500"
-                  href="https://hci.edu.sg"
-                  target="_blank"
-                >
-                  Hwa Chong Institution
-                </Link>
-                , and I am {new Date().getFullYear() - 2010} years old this
-                year! 🎉
+                I am <span className="text-amber-500">Wan Zhi Kai</span>, a
+                Singaporean software engineer, specialised in frontend
+                development. I am currently a high school student at{" "}
+                <span className="text-amber-500">Hwa Chong Instituion</span>,
+                and I am {new Date().getFullYear() - 2010} years old this year!
+                🎉
               </p>{" "}
               <p>
                 I have self-taught programming since I was 5. I am passionate
                 about building{" "}
-                <span className="text-blue-500">
+                <span className="text-amber-500">
                   beautiful, functional, and intuitive
                 </span>{" "}
                 web applications.
               </p>
               <p>
-                I mainly use{" "}
-                <Link
-                  className="text-red-500"
-                  href="https://www.typescriptlang.org/"
-                  target="_blank"
-                >
-                  TypeScript
-                </Link>
-                ,{" "}
-                <Link
-                  className="text-red-500"
-                  href="https://reactjs.org/"
-                  target="_blank"
-                >
-                  ReactJS
-                </Link>
-                ,{" "}
-                <Link
-                  className="text-red-500"
-                  href="https://nextjs.org/"
-                  target="_blank"
-                >
-                  NextJS
-                </Link>
-                , and{" "}
-                <Link
-                  className="text-red-500"
-                  href="https://tailwindcss.com/"
-                  target="_blank"
-                >
-                  TailwindCSS
-                </Link>{" "}
+                I mainly use technologies like{" "}
+                <span className="text-amber-500">
+                  TypeScript, ReactJS, NextJS, and TailwindCSS
+                </span>{" "}
                 in my projects, including this website! 🚀
               </p>
               <p>
                 I am also fluent in{" "}
-                <span className="text-blue-500">English</span>,{" "}
-                <span className="text-blue-500">Chinese</span>, and am learning{" "}
-                <span className="text-blue-500">
+                <span className="text-amber-500">English</span>,{" "}
+                <span className="text-amber-500">Chinese</span>, and am learning{" "}
+                <span className="text-amber-500">
                   Bahasa Indonesia (Indonesian)
                 </span>
                 !
               </p>
-              <Link className="button" href="/story">
-                Read more about my story 📖
+              <Link className="button uppercase !font-light" href="/story">
+                read more
               </Link>
             </div>
           </div>