+++ components/sections/Experience.tsx
@@ -11,11 +11,15 @@ const Experience = () => {
         <SectionHeader
           sequence={["what have i done in work?", 4000, "我的工作经验", 4000]}
         >
-          my work <span className="text-purple-500">experience</span>
+          my work <span className="text-green-500">experience</span>
         </SectionHeader>
-        {/* TODO: remove when got more jobs */}
-        <div className="flex flex-row items-center justify-center space-x-6">
-          {/* <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4 place-items-center"> */}
+        <div
+          className={`grid grid-cols-1 ${
+            experiences.length <= 1
+              ? "xl:grid-cols-1"
+              : "xl:grid-cols-2"
+          } gap-4`}
+        >
           {experiences.map((exp, index) => (
             <ExperienceCard exp={exp} key={index} />
           ))}