+++ components/sections/Projects.tsx
@@ -12,7 +12,7 @@ const Projects = () => {
         <SectionHeader
           sequence={["my project showcase", 4000, "作品展示", 4000]}
         >
-          what have i <span className="text-green-500">built</span>?
+          what have i <span className="text-blue-500">built</span>?
         </SectionHeader>
         <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4 place-items-center">
           {projects.slice(0, 3).map((proj, index) => (
@@ -21,7 +21,7 @@ const Projects = () => {
         </div>
         <div className="flex flex-row w-full justify-end items-center">
           <Link href="/project" className="flex flex-row space-x-4">
-            <button className="button">View All</button>
+            <button className="button uppercase !font-light">view all</button>
           </Link>
         </div>
       </div>