+++ CHANGELOG.md
@@ -1,13 +1,15 @@
-v0.1.0a
+# zk-website CHANGELOG
+
+### v0.1.0a
 - improved contact
 - changed social icons to backtotop
 
-v0.1.1
+### v0.1.1
 - made backtotop show on mobile
 - allow mobile users see hero background video
 - added story page, and moved hobbies into the story page
 
-v0.1.2 - 9/4/2024
+### v0.1.2 - 9/4/2024
 - fixed bugs in story page
 - truncated projects and added a seperate page for project listing
 - each project page now has animation for links to github and live site
@@ -16,13 +18,13 @@ v0.1.2 - 9/4/2024
 - update to all the cards where animation is changed
 - update to work experiences to make each card clickable, and not just the name of the company
 
-v0.1.2a - 9/4/2024
+### v0.1.2a - 9/4/2024
 - added "source code" and version number link in footer
 - forgot to change the card animation for contact
 - forgot to add experience to the footer
 - added links to each skillset
 
-v0.1.3 - 10/4/2024
+### v0.1.3 - 10/4/2024
 - removed hero video on mobile to load faster
 - removed "unoptimized" from images to make the site load faster, especially on mobile (except for the project page where the image does not load unless "unoptimized" is added)
 - fixed padding issues on mobile for programming skillset
@@ -33,3 +35,18 @@ v0.1.3 - 10/4/2024
 - add discord link to contact methods
 - fixed weird-looking header on mobile
 - edited logo to match colors
+
+### v0.1.4 - 11/4/2024
+- changed experience card
+- added new libraries skillset
+- allow 5 cards per row on xl for skillset
+- remove rounded to make it look sharper
+- added more projects
+- edited video for hero
+- removed header gradient
+- changed size of each section
+- skillset is now a separate section
+- removed all other highlight colors, and only kept the amber-500 color
+- made about section larger
+- button redesign
+